move compliance reports to its own page
